Separate contexts into files
